Move common extensions and methods to internal/Utils.kt
